Make clippy happy

+++ roaring/src/bitmap/iter.rs
@@ -732,6 +732,7 @@ impl Extend<u32> for RoaringBitmap {
     /// assert!(rb.contains(1508));
     /// assert!(!rb.contains(5));
     /// ```
+    #[inline]
     fn extend<I: IntoIterator<Item = u32>>(&mut self, values: I) {
         let mut values = values.into_iter();
         let value = match values.next() {
@@ -742,18 +743,18 @@ impl Extend<u32> for RoaringBitmap {
         let (mut currenthb, lowbit) = util::split(value);
         let mut current_container_index = self.find_container_by_key(currenthb);
         let mut current_cont = &mut self.containers[current_container_index];
-        current_cont.insert(lowbit) as u64;
+        current_cont.insert(lowbit);
 
         for val in values {
             let (newhb, lowbit) = util::split(val);
             if currenthb == newhb {
                 // easy case, this could be quite frequent
-                current_cont.insert(lowbit) as u64;
+                current_cont.insert(lowbit);
             } else {
                 currenthb = newhb;
                 current_container_index = self.find_container_by_key(currenthb);
                 current_cont = &mut self.containers[current_container_index];
-                current_cont.insert(lowbit) as u64;
+                current_cont.insert(lowbit);
             }
         }
     }
@@ -777,6 +778,7 @@ impl<'a> Extend<&'a u32> for RoaringBitmap {
     /// assert!(rb.contains(1508));
     /// assert!(!rb.contains(5));
     /// ```
+    #[inline]
     fn extend<I: IntoIterator<Item = &'a u32>>(&mut self, values: I) {
         self.extend(values.into_iter().copied());
     }
